+++ MassParticle/mpRendererOpenGL.cpp
@@ -17,16 +17,16 @@
 class mpRendererOpenGL : public mpRenderer
 {
 public:
-	mpRendererOpenGL();
-	~mpRendererOpenGL();
-	virtual void updateDataTexture(void *tex, const void *data, size_t data_size);
+    mpRendererOpenGL();
+    ~mpRendererOpenGL();
+    virtual void updateDataTexture(void *tex, const void *data, size_t data_size);
 
 private:
 };
 
 mpRenderer* mpCreateRendererOpenGL(void *device)
 {
-	return new mpRendererOpenGL();
+    return new mpRendererOpenGL();
 }
 
 mpRendererOpenGL::mpRendererOpenGL()
@@ -39,13 +39,13 @@ mpRendererOpenGL::~mpRendererOpenGL()
 
 void mpRendererOpenGL::updateDataTexture(void *tex, const void *data, size_t data_size)
 {
-	const int num_texels = data_size / 16;
-	ivec2 pos = ivec2(0, 0);
-	ivec2 size = ivec2(mpDataTextureWidth, ceildiv(num_texels, mpDataTextureWidth));
+    const size_t num_texels = data_size / 16;
+    ivec2 pos = ivec2(0, 0);
+    ivec2 size = ivec2(mpDataTextureWidth, ceildiv((int)num_texels, mpDataTextureWidth));
 
-	glBindTexture(GL_TEXTURE_2D, (GLuint)(size_t)tex);
-	glTexSubImage2D(GL_TEXTURE_2D, 0, pos.x, pos.y, size.x, size.y, GL_RGBA, GL_FLOAT, data);
-	glBindTexture(GL_TEXTURE_2D, 0);
+    glBindTexture(GL_TEXTURE_2D, (GLuint)(size_t)tex);
+    glTexSubImage2D(GL_TEXTURE_2D, 0, pos.x, pos.y, size.x, size.y, GL_RGBA, GL_FLOAT, data);
+    glBindTexture(GL_TEXTURE_2D, 0);
 }
 
 #endif // SUPPORT_OPENGL
